requirements.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
idna==2.10
joblib==1.0.0
Markdown==3.3.3
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.5.1.48
Pillow==7.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
torch==1.7.1
torch-dct==0.1.5
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
